+++ sw.js
@@ -1,5 +1,5 @@
 const baseCacheName = 'Formuliser';
-const versionName = 'v6_f3';
+const versionName = 'v7_dev1';
 const cacheName = baseCacheName + '-' + versionName;
 const runtimeCacheName = cacheName + '-runtime';